FIX: Compile issues on Windows
The warning levels are more picky on Windows. Fix all small
issues so it's buildable on Windows again.

+++ assets/objects/robots/BaseBot.h
@@ -81,7 +81,7 @@ public:
     void setTimeToTopSpeedCallback(std::function<void(float)> onTimeToTopSpeedChanged);
     void setTopSpeedAccelerationCallback(std::function<void(float)> onTopSpeedAccelerationChanged);
     void setTopAccelerationCallback(std::function<void(float)> onTopAccelerationChanged);
-    void setTimeMovingCallback(std::function<void(float)> onTimeMovingChanged);
+    void setTimeMovingCallback(std::function<void(unsigned int)> onTimeMovingChanged);
     void enableMotor(BaseBot::WheelMotorIndex wheelMotorIndex);
     void disableMotor(BaseBot::WheelMotorIndex wheelMotorIndex);
     float getAngularDamping() const;
@@ -103,9 +103,9 @@ private:
     std::function<void(float)> m_onTimeToTopSpeedChanged = nullptr;
     std::function<void(float)> m_onTopSpeedAccelerationChanged = nullptr;
     std::function<void(float)> m_onTopAccelerationChanged = nullptr;
-    std::function<void(float)> m_onTimeMovingChanged = nullptr;
-    float m_lastFwdStandStillTime = 0.0f;
-    float m_lastRotStandStillTime = 0.0f;
+    std::function<void(unsigned int)> m_onTimeMovingChanged = nullptr;
+    unsigned int m_lastFwdStandStillTime = 0;
+    unsigned int m_lastRotStandStillTime = 0;
     float m_recordedTopSpeed = 0.0f;
     float m_recordedTopAcceleration = 0.0f;
     float m_timeToReachTopSpeed = 0.0f;
