+++ src/scene/Scene.cpp
@@ -94,11 +94,11 @@ void Scene::addMenu(ImGuiMenu *menu)
 unsigned int Scene::getSecondsSinceStart() const
 {
     const auto timeNow = std::chrono::system_clock::now();
-    return std::chrono::duration<double>(timeNow - m_startTime).count();
+    return static_cast<unsigned int>(std::chrono::duration<double>(timeNow - m_startTime).count());
 }
 
 unsigned int Scene::getMillisecondsSinceStart() const
 {
     const auto timeNow = std::chrono::system_clock::now();
-    return std::chrono::duration<double, std::milli>(timeNow - m_startTime).count();
+    return static_cast<unsigned int>(std::chrono::duration<double, std::milli>(timeNow - m_startTime).count());
 }
